Actualizacion de ejercicios. Agregado de ejercicios clase 5 y algunos ejercicios de POO

+++ EjercitacionClase2D-LaplaceJulieta/Clase 03/C3-Ej-A01/Program.cs
@@ -6,24 +6,24 @@ namespace C3_Ej_A01
     {
         static void Main(string[] args)
         {
-            int numeroIngresado;
+            double numeroIngresado;
             bool numeroIngresadoValido;
-            int resultadoFactorial;
+            double resultadoFactorial;
 
             Console.WriteLine("Ingrese un numero mayor a 0 para saber su factorial:");
-            numeroIngresadoValido = int.TryParse(Console.ReadLine(), out numeroIngresado);
+            numeroIngresadoValido = double.TryParse(Console.ReadLine(), out numeroIngresado);
             while (!numeroIngresadoValido || numeroIngresado <= 0)
             {
                 Console.WriteLine("Dato Invalido. Ingrese un numero mayor a 0 para saber su factorial:");
-                numeroIngresadoValido = int.TryParse(Console.ReadLine(), out numeroIngresado);
+                numeroIngresadoValido = double.TryParse(Console.ReadLine(), out numeroIngresado);
             } 
 
             resultadoFactorial = CalcularFactorial(numeroIngresado);
             Console.WriteLine($"El factorial de {numeroIngresado} es {resultadoFactorial} ");
         }
-        public static int CalcularFactorial(int numero)
+        public static double CalcularFactorial(double numero)
         {
-            int factorial = 1;
+            double factorial = 1;
             for (int i = 1; i <= numero; i++)
             {
                 factorial= factorial*i;
